Overhaul and Cleanup of the README File (#250)
This is meant to add more information as well as clarify current
information. Also want to align with the [README of the Java version of
Attachments](https://github.com/cap-java/cds-feature-attachments/blob/main/README.md).
Steps remaining:
- [x] Finalize structure and sections 
   - [x] Removing Java-specific sections and information
- [x] Confirm and clarify details
- [x] Add in links to sources


This initial change should make the rest of the editing process more
streamlined

---------

Co-authored-by: Simon Kobler <32038731+KoblerS@users.noreply.github.com>
Co-authored-by: Lisa Julia Nebel <lisajuliafog@yahoo.de>
Co-authored-by: René Jeglinsky <rene.jeglinsky@sap.com>
Co-authored-by: Simon Kobler <simon.kobler@sap.com>