Leverage generic handlers to allow for extension and feature toggles (#292)
Register handlers in a generic fashion so the plugin can be put behind a
feature toggle or extension.

The scenario is that `cds.model` only has the base model but that might
not contain attachments. This currently leads to attachment handlers not
being registered and then once a feature toggle is activated the
handlers are missing to handle the added attachment.

Using generic handlers fixes this.

---------

Co-authored-by: Simon Kobler <32038731+KoblerS@users.noreply.github.com>