Fix separate object store race condition and add standard plan (#303)
When using MTX + separate Object Store mode there is a chance for a race
condition because the Object store bucket is stored as a property on the
service which is however shared across all tenants.

The PR addresses this by introducing getBucket / getClient methods and
within these the tenant-specific cache lookup is performed / a new
client is created.

---------

Co-authored-by: Simon Kobler <32038731+KoblerS@users.noreply.github.com>